# docs/swagger.yaml
...
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.ConnectorInfoResp:
    properties:
      icon:
        type: string
      link:
        type: string
      name:
        type: string
    type: object
  schema.ConnectorUserInfoResp:
    properties:
      binding:
        type: boolean
      external_id:
        type: string
      icon:
        type: string
      link:
        type: string
      name:
        type: string
    type: object
  schema.ExternalLoginBindingUserSendEmailReq:
    properties:
      binding_key:
        maxLength: 100
        type: string
      email:
        maxLength: 512
        type: string
      must:
        description: |-
          If must is true, whatever email if exists, try to bind user.
          If must is false, when email exist, will only be prompted with a warning.
        type: boolean
    required:
    - binding_key
    - email
    type: object
  schema.ExternalLoginBindingUserSendEmailResp:
    properties:
      access_token:
        type: string
      email_exist_and_must_be_confirmed:
        type: boolean
    type: object
  schema.ExternalLoginUnbindingReq:
    properties:
      external_id:
        maxLength: 128
        type: string
    required:
    - external_id
    type: object
...
info:
  contact: {}
paths:
...
  /answer/api/v1/connector/binding/email:
    post:
      consumes:
      - application/json
      description: external login binding user send email
      parameters:
      - description: external login binding user send email
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ExternalLoginBindingUserSendEmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ExternalLoginBindingUserSendEmailResp'
              type: object
      summary: external login binding user send email
      tags:
      - PluginConnector
  /answer/api/v1/connector/info:
    get:
      description: get all enabled connectors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.ConnectorInfoResp'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: get all enabled connectors
      tags:
      - PluginConnector
  /answer/api/v1/connector/user/info:
    get:
      description: get all connectors info about user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.ConnectorUserInfoResp'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: get all connectors info about user
      tags:
      - PluginConnector
  /answer/api/v1/connector/user/unbinding:
    delete:
      consumes:
      - application/json
      description: unbind external user login
      parameters:
      - description: ExternalLoginUnbindingReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ExternalLoginUnbindingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: unbind external user login
      tags:
      - PluginConnector
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"